Adds integration page basic styling
