+++ package.json
@@ -172,6 +172,7 @@
 		"react-error-boundary": "^3.1.4",
 		"react-hot-toast": "^2.2.0",
 		"react-intersection-observer": "^8.33.1",
+		"react-markdown": "^8.0.3",
 		"react-player": "^2.9.0",
 		"rivet-graphql": "^0.3.1",
 		"search-insights": "^1.8.0",
