+++ src/views/product-integrations-landing/components/integrations-list/style.module.css
@@ -6,25 +6,6 @@
 	}
 }
 
-.badge {
-	composes: hds-typography-body-100 from global;
-	color: var(--token-color-foreground-faint);
-	font-weight: var(--token-typography-font-weight-medium);
-
-	/* TODO, hardcoded to vault's color, should be
-	 * based off of product color */
-	background: var(--badge-color);
-	padding: 2px 8px;
-	border-radius: 30px;
-	display: flex;
-	align-items: center;
-
-	& > svg {
-		height: 14px;
-		margin-right: 2px;
-	}
-}
-
 .heading {
 	composes: hds-typography-body-200 from global;
 	color: var(--token-color-foreground-strong);
