+++ src/views/product-integrations-landing/components/tier-badge/style.module.css
@@ -1,0 +1,28 @@
+.tierBadge {
+	composes: hds-typography-body-100 from global;
+	color: var(--token-color-foreground-faint);
+	font-weight: var(--token-typography-font-weight-medium);
+	background: var(--badge-color);
+	padding: 2px 8px;
+	border-radius: 30px;
+	display: flex;
+	align-items: center;
+
+	& > svg {
+		height: 14px;
+		margin-right: 2px;
+	}
+
+	&.large {
+		font-size: 16px;
+		padding: 4px 16px;
+		font-family: var(--token-typography-display-500-font-family);
+		line-height: var(--token-typography-display-500-line-height);
+
+		& > svg {
+			height: 18px;
+			width: auto;
+			margin-right: 4px;
+		}
+	}
+}
